standardize caching of properties of ADORecordSet class #747
The method of retrieval of the field properties of a recordset now has a standardized method, setFieldObjectsCache(), which is called immediately after initial recordset retrieval. The objects retrieved are stored in a parent level variable, fieldObjectsCache.  calls to fetchfield() now simply reads the cache instead of re-querying the database.

This methodology is usable for all drivers